healthcheck fixes

+++ workstation/kira/monitor-snapshot.sh
@@ -16,17 +16,15 @@ SNAP_PROGRESS="$SNAP_STATUS/progress"
 SNAP_DONE="$SNAP_STATUS/done"
 SNAP_LATEST="$SNAP_STATUS/latest"
 
+while [ ! -f $SCAN_DONE ] ; do
+    echo "INFO: Waiting for monitor scan to finalize run..."
+    sleep 10
+done
 
-if [ ! -f $SCAN_DONE ]; then
-    echo "INFO: Scan is not done yet, aborting snapshot monitor"
-    sleep 60
-    exit 0
-fi
+LATEST_BLOCK=$(tryCat $LATEST_BLOCK_SCAN_PATH "0")
 
 INTERX_REFERENCE_DIR="$DOCKER_COMMON/interx/cache/reference"
 INTERX_SNAPSHOT_PATH="$INTERX_REFERENCE_DIR/snapshot.zip"
-LATEST_BLOCK=$(tryCat $LATEST_BLOCK_SCAN_PATH "0")
-DOCKER_SNAP_DESTINATION="$DOCKER_COMMON_RO/snap.zip"
 
 set +x
 echoWarn "------------------------------------------------"
@@ -47,6 +45,7 @@ if [ -f "$SNAP_LATEST" ] && [ -f "$SNAP_DONE" ]; then
     if [ -f "$SNAP_LATEST_FILE" ] && [ "$KIRA_SNAP_PATH" != "$SNAP_LATEST_FILE" ]; then
         KIRA_SNAP_PATH=$SNAP_LATEST_FILE
         CDHelper text lineswap --insert="KIRA_SNAP_PATH=\"$KIRA_SNAP_PATH\"" --prefix="KIRA_SNAP_PATH=" --path=$ETC_PROFILE --append-if-found-not=True
+        $KIRA_SCRIPTS/container-delete.sh "snapshot" || echoErrr "ERROR: Failed to delete snapshot container"
         CHECKSUM_TEST="true"
     fi
 fi
@@ -57,20 +56,18 @@ if [ "${CHECKSUM_TEST,,}" == "true" ] || ( [ -f "$KIRA_SNAP_PATH" ] && [ -z "$KI
     CDHelper text lineswap --insert="KIRA_SNAP_SHA256=\"$KIRA_SNAP_SHA256\"" --prefix="KIRA_SNAP_SHA256=" --path=$ETC_PROFILE --append-if-found-not=True
 fi
 
-if [ -f "$KIRA_SNAP_PATH" ] && [ "${SNAP_EXPOSE,,}" == "true" ] && [ "$KIRA_SNAP_SHA256" != "$INTERX_SNAP_SHA256" ] ; then
-    if [ "$KIRA_SNAP_SHA256" != "$INTERX_SNAP_SHA256" ]; then
-        echoInfo "INFO: Latest snapshot is NOT exposed yet"
-        mkdir -p $INTERX_REFERENCE_DIR
-        SUCCESS="true" && ln -fv "$KIRA_SNAP_PATH" "$INTERX_SNAPSHOT_PATH" || SUCCESS="false"
-        [ "${SUCCESS,,}" == "true" ] && ln -fv "$KIRA_SNAP_PATH" "$DOCKER_SNAP_DESTINATION" || SUCCESS="false"
-        if [ "${SUCCESS,,}" == "true" ] ; then
-            CDHelper text lineswap --insert="INTERX_SNAP_SHA256=\"$KIRA_SNAP_SHA256\"" --prefix="INTERX_SNAP_SHA256=" --path=$ETC_PROFILE --append-if-found-not=True
-        else
-            rm -fv "$INTERX_SNAPSHOT_PATH" "$DOCKER_SNAP_DESTINATION" || echoErr "ERROR: Failed to cleanup symlinks"
-            echoErr "ERROR: Failed to create one of the symlinks $KIRA_SNAP_PATH => $INTERX_SNAPSHOT_PATH, $KIRA_SNAP_PATH => $DOCKER_SNAP_DESTINATION"
-        fi
-    else
-        echoInfo "INFO: Latest snapshot was already exposed, no need for updates"
+if [ -f "$KIRA_SNAP_PATH" ] && [ "$KIRA_SNAP_SHA256" != "$INTERX_SNAP_SHA256" ] ; then
+    echoInfo "INFO: Latest snapshot is NOT exposed yet"
+    mkdir -p $INTERX_REFERENCE_DIR
+
+    if [ "${SNAP_EXPOSE,,}" == "true" ]; then
+        ln -fv "$KIRA_SNAP_PATH" "$INTERX_SNAPSHOT_PATH"
+        echoInfo "INFO: Symlink $KIRA_SNAP_PATH => $INTERX_SNAPSHOT_PATH was created"
+        CDHelper text lineswap --insert="INTERX_SNAP_SHA256=\"$KIRA_SNAP_SHA256\"" --prefix="INTERX_SNAP_SHA256=" --path=$ETC_PROFILE --append-if-found-not=True
+    elif [ "${SNAP_EXPOSE,,}" == "false" ] && [ -f "$INTERX_SNAPSHOT_PATH" ] ; then
+        rm -fv "$INTERX_SNAPSHOT_PATH"
+        CDHelper text lineswap --insert="INTERX_SNAP_SHA256=\"\"" --prefix="INTERX_SNAP_SHA256=" --path=$ETC_PROFILE --append-if-found-not=True
+        echoInfo "INFO: Symlink $KIRA_SNAP_PATH => $INTERX_SNAPSHOT_PATH was removed"
     fi
 elif [ -f "$INTERX_SNAPSHOT_PATH" ] && ([ "${SNAP_EXPOSE,,}" == "false" ] || [ -z "$KIRA_SNAP_PATH" ]); then
     echoInfo "INFO: Removing publicly exposed snapshot..."
@@ -88,10 +85,10 @@ fi
 if [ -z "$AUTO_BACKUP_EXECUTED_TIME" ] ; then
     echoInfo "INFO: Backup was never scheaduled before, it will be set to be executed within 1 interval from current time"
     CDHelper text lineswap --insert="AUTO_BACKUP_EXECUTED_TIME=\"$(date -u +%s)\"" --prefix="AUTO_BACKUP_EXECUTED_TIME=" --path=$ETC_PROFILE --append-if-found-not=True
-elif [ -f $SCAN_DONE ] && [ "${AUTO_BACKUP_ENABLED,,}" == "true" ] && [ $LATEST_BLOCK -gt $AUTO_BACKUP_LAST_BLOCK ]; then
+elif [ -f $SCAN_DONE ] && [ "${AUTO_BACKUP_ENABLED,,}" == "true" ] && [ $LATEST_BLOCK -gt $AUTO_BACKUP_LAST_BLOCK ] && [[ $MAX_SNAPS -gt 0 ]]; then
     ELAPSED_TIME=$(($(date -u +%s) - $AUTO_BACKUP_EXECUTED_TIME))
     INTERVAL_AS_SECOND=$(($AUTO_BACKUP_INTERVAL * 3600))
-    if [[ $ELAPSED_TIME -gt $INTERVAL_AS_SECOND ]]; then
+    if [[ $ELAPSED_TIME -gt $INTERVAL_AS_SECOND ]] ; then
         rm -fv $SCAN_DONE
         [ -f "$KIRA_SNAP_PATH" ] && SNAP_PATH_TMP=$KIRA_SNAP_PATH || SNAP_PATH_TMP=""
         $KIRA_MANAGER/containers/start-snapshot.sh "$LATEST_BLOCK" "$SNAP_PATH_TMP" &> "${SNAPSHOT_SCAN_PATH}-start.log"
